(arm64) More fixups

+++ src/sccz80/expr.c
@@ -507,7 +507,7 @@ int heira(LVALUE *lval)
             rvalue(lval);
         intcheck(lval, lval);
         com(lval);
-        lval->const_val = (int64_t)~(uint64_t)lval->const_val;
+        lval->const_val = ~(uint64_t)lval->const_val;
         lval->stage_add = NULL;
         return 0;
     } else if (cmatch('!')) {
